Move InstallationCandidate to own module
Move InstallationCandidate to `pip.models.installationcandidate`

because it's only used by `pip.index` and having it live inside
`pip.req`, made `pip.index` depend on `pip.req`, thereby creating a
circular dependency.

By moving `InstallationCandidate` from `pip.req.req_requirement`, we
can:

1. Delete the somewhat confusingly named `pip.req.req_requirement`
   altogether, as it only contained this one class.

2. Eliminate imports inside functions in `pip.req.req_install`, which
   were necessary because of the circular dependency, which is now gone.
